+++ src/Vsix/Completion/CompletionController.cs
@@ -67,15 +67,10 @@ namespace EditorConfig
                         case VSConstants.VSStd2KCmdID.TYPECHAR:
                             char ch = GetTypeChar(pvaIn);
 
-                            if (ch == ':')
+                            if (ch == ':' || ch == '=')
                             {
                                 Cancel();
                                 StartSession();
-                            }
-                            else if (ch == '=')
-                            {
-                                Complete(true);
-                                hresult = Next.Exec(pguidCmdGroup, nCmdID, nCmdexecopt, pvaIn, pvaOut);
                             }
                             else if (!char.IsPunctuation(ch) && !char.IsControl(ch))
                             {